Conflicts:
	src/library_glfw.js

+++ system/lib/libc/musl/src/thread/pthread_rwlock_timedwrlock.c
@@ -5,7 +5,7 @@ int pthread_rwlock_timedwrlock(pthread_rwlock_t *restrict rw, const struct times
 #ifdef __EMSCRIPTEN__
 	/// XXX Emscripten: The spec allows detecting when multiple write locks would deadlock, which we do here to avoid hangs.
 	/// If attempting to lock the write lock that we already own, error out.
-	if (rw->_rw_wr_owner == pthread_self()) return EDEADLK;
+	if (rw->_rw_wr_owner == (int)pthread_self()) return EDEADLK;
 #endif
 	int r, t;
 	while ((r=pthread_rwlock_trywrlock(rw))==EBUSY) {
@@ -20,7 +20,7 @@ int pthread_rwlock_timedwrlock(pthread_rwlock_t *restrict rw, const struct times
 #ifdef __EMSCRIPTEN__
 	/// XXX Emscripten: The spec allows detecting when multiple write locks would deadlock, which we do here to avoid hangs.
 	/// Mark this thread as the owner of this write lock.
-	rw->_rw_wr_owner = pthread_self();
+	rw->_rw_wr_owner = (int)pthread_self();
 #endif
 	return r;
 }
